Improve HTTP response handling in site.c
Refactor string tokenization and condition checks in the
`thread_fn` function to enhance parsing accuracy and
compliance with HTTP standards. Update the initialization
of the response buffer to prevent garbage values. Change
`header.range_end` to an int type for better range handling,
and refine the calculation of `content_length`. Add
support for additional headers like "Accept-Ranges" and
"Content-Disposition" for multimedia files. Extend
`getcontenttype` to handle more video and audio MIME types,
and modify `getcontentrange` to return -1 for unspecified
end ranges, improving overall robustness and functionality.

+++ site.h
@@ -29,8 +29,8 @@ typedef struct Header {
   char path[BUFFER_SIZE];
   bool keep_alive;
   bool range_request;
-  size_t range_start;
-  size_t range_end;
+  int range_start;
+  int range_end;
   List* headers;
 } Header;
